FinalProject/requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
mysqlclient==2.1.1
python-dotenv==1.0.0
Pillow==10.0.0
